+++ TableViewExtensions.m
@@ -83,8 +83,8 @@
  */
 -(NSString *)view:(NSView *)view stringForToolTip:(NSToolTipTag)tag point:(NSPoint)point userData:(void *)matrix
 {
-	int rowIndex = [self rowAtPoint:point];
-	int columnIndex = [self columnAtPoint:point];
+	NSInteger rowIndex = [self rowAtPoint:point];
+	NSInteger columnIndex = [self columnAtPoint:point];
 	NSTableColumn *tableColumn = (columnIndex != -1) ? [[self tableColumns] objectAtIndex:columnIndex] : nil;
 	return (columnIndex != -1) ? [[self delegate] tableView:self toolTipForTableColumn:tableColumn row:rowIndex] : @"";
 }
